Update dnsconf.yml
ok so this part took me a while to figure out but this is the best way to describe it:

the first part is the resolve conf part. this should resolve any issues that might arise in the config. it also defines the system we are using which is CentOS. 

resolv.conf should be the file that houses all the actual config portions for the DNS which i suppose (at the time of writing this) i should change the previous file name to.

but the second part should ask you for the dns info as a prompt and then commit that to the dhclient file.