Update DNS Long
a note when i was searching the web for ideas. 
ansible needs to logon from the control computer to the managed computers; this means that the managed computers already have their networking setup, so ansible would normally not be used for setting up networking. Since we are new to ansible, I would recommend, if possible, to postpone investigating this issue and come back to it when we have some more experience with ansible